NDHR-39 라우팅 구조 수정
enter 삭제, 기존 enter를 index로.
기존 index는 새롭게 timeline 라우팅으로 보냄. (모아보는)
